v0 belief

+++ hanabi_lib/canonical_encoders.h
@@ -33,7 +33,7 @@ class CanonicalObservationEncoder : public ObservationEncoder {
       : parent_game_(parent_game) {}
 
   std::vector<int> Shape() const override;
-  std::vector<int> Encode(const HanabiObservation& obs,
+  std::vector<float> Encode(const HanabiObservation& obs,
                           bool show_own_cards=false) const override;
 
   ObservationEncoder::Type type() const override {
